move test to cli test file
